fix(e2e): isolate space-sub-routes tests with unique workspace dirs
Multiple E2E test files run in parallel and all created spaces using the
same workspace root path. Since the `spaces` table has a UNIQUE constraint
on `workspace_path`, parallel tests raced and hit "A space already exists"
errors.

Fix by adding `createUniqueSpaceDir()` helper to `space-helpers.ts` that
creates a unique subdirectory per test run. The `space-sub-routes.e2e.ts`
now uses this helper so its spaces get isolated workspace paths that
don't conflict with concurrent tests.
