+++ packages/e2e/tests/features/space-sub-routes.e2e.ts
@@ -17,11 +17,37 @@ import { test, expect } from '../../fixtures';
 import { waitForWebSocketConnected, getWorkspaceRoot } from '../helpers/wait-helpers';
 import {
 	createSpaceViaRpc,
-	createSpaceTaskViaRpc,
+	createUniqueSpaceDir,
 	deleteSpaceViaRpc,
 } from '../helpers/space-helpers';
 
 const DESKTOP_VIEWPORT = { width: 1280, height: 720 };
+
+/**
+ * Create a task via RPC. For use in beforeEach setup only.
+ * Returns the new task's id.
+ */
+async function createTaskViaRpc(
+	page: Parameters<typeof waitForWebSocketConnected>[0],
+	spaceId: string,
+	title: string
+): Promise<string> {
+	const id = await page.evaluate(
+		async ({ spaceId, title }) => {
+			const hub = window.__messageHub || window.appState?.messageHub;
+			if (!hub?.request) throw new Error('MessageHub not available');
+			const task = (await hub.request('spaceTask.create', {
+				spaceId,
+				title,
+				description: '',
+			})) as { id: string };
+			return task.id;
+		},
+		{ spaceId, title }
+	);
+	if (!id) throw new Error('spaceTask.create returned no id');
+	return id;
+}
 
 /**
  * Create a standalone session via RPC. For use in beforeEach setup only.
@@ -75,9 +101,12 @@ test.describe('Space Sub-Routes Deep Links', () => {
 		await waitForWebSocketConnected(page);
 
 		const workspaceRoot = await getWorkspaceRoot(page);
+		// Use a unique subdirectory per test to avoid conflicts with other parallel tests
+		// that also create spaces (workspace_path has a UNIQUE constraint in the DB).
+		const spaceWorkspacePath = createUniqueSpaceDir(workspaceRoot, 'sub-routes');
 		const spaceName = `E2E Sub-Routes Test ${Date.now()}`;
-		spaceId = await createSpaceViaRpc(page, workspaceRoot, spaceName);
-		taskId = await createSpaceTaskViaRpc(page, spaceId, `Test Task ${Date.now()}`);
+		spaceId = await createSpaceViaRpc(page, spaceWorkspacePath, spaceName);
+		taskId = await createTaskViaRpc(page, spaceId, `Test Task ${Date.now()}`);
 		sessionId = await createSessionViaRpc(page, workspaceRoot);
 	});
