+++ packages/e2e/tests/helpers/space-helpers.ts
@@ -5,6 +5,8 @@
  * All test actions and assertions must go through the browser UI.
  */
 
+import { mkdirSync } from 'node:fs';
+import { join } from 'node:path';
 import type { Page } from '@playwright/test';
 
 /**
@@ -53,30 +55,27 @@ export async function deleteSpaceViaRpc(page: Page, spaceId: string): Promise<vo
 }
 
 /**
- * Create a standalone task in a space via RPC. For use in test setup only.
- * Returns the new task's id.
+ * Create a unique workspace subdirectory for a space test.
+ *
+ * Multiple E2E tests run in parallel and all share the same workspace root.
+ * Since the `spaces` table has a UNIQUE constraint on `workspace_path`, parallel
+ * tests that all try to create spaces at the workspace root will race and conflict.
+ *
+ * This helper creates a unique subdirectory within the workspace root so each
+ * test gets its own isolated path. The directory is created synchronously (Node.js
+ * side) before the space is created via RPC.
+ *
+ * @param workspaceRoot - The base workspace root (from `getWorkspaceRoot(page)`)
+ * @param prefix - Optional prefix for the subdirectory name (for easier debugging)
+ * @returns The unique subdirectory path
  */
-export async function createSpaceTaskViaRpc(
-	page: Page,
-	spaceId: string,
-	title: string,
-	description = ''
-): Promise<string> {
-	const id = await page.evaluate(
-		async ({ spaceId, title, description }) => {
-			const hub = window.__messageHub || window.appState?.messageHub;
-			if (!hub?.request) throw new Error('MessageHub not available');
-			const task = (await hub.request('spaceTask.create', {
-				spaceId,
-				title,
-				description,
-			})) as { id: string };
-			return task.id;
-		},
-		{ spaceId, title, description }
+export function createUniqueSpaceDir(workspaceRoot: string, prefix = 'space'): string {
+	const uniqueDir = join(
+		workspaceRoot,
+		`${prefix}-${Date.now()}-${Math.random().toString(36).slice(2, 9)}`
 	);
-	if (!id) throw new Error('spaceTask.create returned no id');
-	return id;
+	mkdirSync(uniqueDir, { recursive: true });
+	return uniqueDir;
 }
 
 /**
